+++ source/D2Game/src/UNIT/Party.cpp
@@ -72,7 +72,7 @@ int16_t __fastcall sub_6FCB9C40(D2GameStrc* pGame)
 
     D2PartyControlStrc* pPartyControl = pGame->pPartyControl;
 
-    int16_t nPartyId = std::max(pPartyControl->field_0, 3i16);
+    int16_t nPartyId = std::max(pPartyControl->field_0, (int16_t)3);
 
     while (1)
     {
@@ -89,7 +89,7 @@ int16_t __fastcall sub_6FCB9C40(D2GameStrc* pGame)
 
         ++nPartyId;
 
-        nPartyId = std::max(nPartyId, 3i16);
+        nPartyId = std::max(nPartyId, (int16_t)3);
     }
 
     pPartyControl->field_0 = nPartyId + 1;
